+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MiniAppMessageBridge.kt
@@ -4,15 +4,22 @@ import android.webkit.JavascriptInterface
 import androidx.annotation.VisibleForTesting
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
+import com.rakuten.tech.mobile.miniapp.MiniAppInfo
 import com.rakuten.tech.mobile.miniapp.display.WebViewListener
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermission
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionCache
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionManager
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionResult
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionResult
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionResult
 
 @Suppress("TooGenericExceptionCaught", "SwallowedException", "TooManyFunctions")
 /** Bridge interface for communicating with mini app. **/
 abstract class MiniAppMessageBridge {
     private lateinit var webViewListener: WebViewListener
+    private lateinit var customPermissionCache: MiniAppCustomPermissionCache
+    private lateinit var miniAppInfo: MiniAppInfo
 
     /** Get provided id of mini app for any purpose. **/
     abstract fun getUniqueId(): String
@@ -23,11 +30,35 @@ abstract class MiniAppMessageBridge {
         callback: (isGranted: Boolean) -> Unit
     )
 
-    /** Post custom permissions request with names and descriptions from external. **/
+    /**
+     * Post custom permissions request.
+     * @param permissionsWithDescription list of name and descriptions of custom permissions sent from external.
+     * @param callback to invoke a list of name and grant results of custom permissions sent from hostapp.
+     */
     abstract fun requestCustomPermissions(
-        permissions: List<Pair<MiniAppCustomPermissionType, String>>,
-        callback: (grantResult: String) -> Unit
+        permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>,
+        callback: (List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>) -> Unit
     )
+
+    /**
+     * Share content info [ShareInfo]. This info is provided by mini app.
+     * @param content The content property of [ShareInfo] object.
+     * @param callback The executed action status should be notified back to mini app.
+     **/
+    abstract fun shareContent(
+        content: String,
+        callback: (isSuccess: Boolean, message: String?) -> Unit
+    )
+
+    internal fun init(
+        webViewListener: WebViewListener,
+        customPermissionCache: MiniAppCustomPermissionCache,
+        miniAppInfo: MiniAppInfo
+    ) {
+        this.webViewListener = webViewListener
+        this.customPermissionCache = customPermissionCache
+        this.miniAppInfo = miniAppInfo
+    }
 
     /** Handle the message from external. **/
     @JavascriptInterface
@@ -38,6 +69,7 @@ abstract class MiniAppMessageBridge {
             ActionType.GET_UNIQUE_ID.action -> onGetUniqueId(callbackObj)
             ActionType.REQUEST_PERMISSION.action -> onRequestPermission(callbackObj)
             ActionType.REQUEST_CUSTOM_PERMISSIONS.action -> onRequestCustomPermissions(jsonStr)
+            ActionType.SHARE_INFO.action -> onShareContent(callbackObj.id, jsonStr)
         }
     }
 
@@ -73,32 +105,32 @@ abstract class MiniAppMessageBridge {
 
         try {
             callbackObj = Gson().fromJson(jsonStr, CustomPermissionCallbackObj::class.java)
-
             val permissionObjList = arrayListOf<CustomPermissionObj>()
             callbackObj.param?.permissions?.forEach {
                 permissionObjList.add(CustomPermissionObj(it.name, it.description))
             }
 
-            val permissionPairList = arrayListOf<Pair<MiniAppCustomPermissionType, String>>()
-            permissionObjList.forEach {
-                MiniAppCustomPermissionType.getValue(it.name)?.let { type ->
-                    permissionPairList.add(Pair(type, it.description))
-                }
-                if (MiniAppCustomPermissionType.getValue(it.name) == null)
-                    permissionPairList.add(
-                        Pair(
-                            MiniAppCustomPermissionType.UNKNOWN,
-                            it.description
-                        )
-                    )
-            }
+            val permissionsWithDescription = MiniAppCustomPermissionManager()
+                .preparePermissionsWithDescription(permissionObjList)
 
             requestCustomPermissions(
-                permissionPairList
-            ) { grantResult ->
+                permissionsWithDescription
+            ) { permissionsWithResult ->
+                // store values in SDK cache
+                val miniAppCustomPermission = MiniAppCustomPermission(
+                    miniAppId = miniAppInfo.id,
+                    pairValues = permissionsWithResult
+                )
+                customPermissionCache.storePermissions(miniAppCustomPermission)
+
+                // send JSON response to miniapp
                 onRequestCustomPermissionsResult(
                     callbackId = callbackObj.id,
-                    grantResult = grantResult
+                    jsonResult = MiniAppCustomPermissionManager().createJsonResponse(
+                        customPermissionCache,
+                        miniAppCustomPermission.miniAppId,
+                        permissionsWithDescription
+                    )
                 )
             }
         } catch (e: Exception) {
@@ -111,6 +143,23 @@ abstract class MiniAppMessageBridge {
         }
     }
 
+    private fun onShareContent(callbackId: String, jsonStr: String) {
+        try {
+            val callbackObj = Gson().fromJson(jsonStr, ShareInfoCallbackObj::class.java)
+
+            shareContent(
+                callbackObj.param.shareInfo.content
+            ) { isSuccess, message ->
+                if (isSuccess)
+                    postValue(callbackId, message ?: SUCCESS)
+                else
+                    postError(callbackId, message ?: "Cannot share content: Unknown error message from hostapp.")
+            }
+        } catch (e: Exception) {
+            postError(callbackId, "Cannot share content: ${e.message}")
+        }
+    }
+
     @VisibleForTesting
     /** Inform the permission request result to MiniApp. **/
     internal fun onRequestPermissionsResult(callbackId: String, isGranted: Boolean) {
@@ -120,10 +169,10 @@ abstract class MiniAppMessageBridge {
             postError(callbackId, MiniAppPermissionResult.getValue(isGranted).type)
     }
 
-    /** Inform the permission request result to MiniApp. **/
+    /** Inform the custom permission request result to MiniApp. **/
     @Suppress("LongMethod", "FunctionMaxLength")
-    internal fun onRequestCustomPermissionsResult(callbackId: String, grantResult: String) {
-        postValue(callbackId, grantResult)
+    internal fun onRequestCustomPermissionsResult(callbackId: String, jsonResult: String) {
+        postValue(callbackId, jsonResult)
     }
 
     /** Return a value to mini app. **/
@@ -134,9 +183,5 @@ abstract class MiniAppMessageBridge {
     /** Emit an error to mini app. **/
     internal fun postError(callbackId: String, errorMessage: String) {
         webViewListener.runErrorCallback(callbackId, errorMessage)
-    }
-
-    internal fun setWebViewListener(webViewListener: WebViewListener) {
-        this.webViewListener = webViewListener
     }
 }