+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/permission/MiniAppCustomPermissionManagerSpec.kt
@@ -1,14 +1,14 @@
 package com.rakuten.tech.mobile.miniapp.permission
 
 import com.nhaarman.mockitokotlin2.*
-import com.rakuten.tech.mobile.miniapp.MiniApp
+import com.rakuten.tech.mobile.miniapp.js.CustomPermissionObj
 import org.junit.Before
 import org.junit.Test
 import kotlin.test.assertEquals
 
 class MiniAppCustomPermissionManagerSpec {
     private lateinit var miniAppCustomPermissionManager: MiniAppCustomPermissionManager
-    private var miniApp: MiniApp = mock()
+    private var miniAppCustomPermissionCache: MiniAppCustomPermissionCache = mock()
     private val miniAppId: String = "miniAppId"
     private val miniAppCustomPermission = MiniAppCustomPermission(
         miniAppId,
@@ -22,20 +22,38 @@ class MiniAppCustomPermissionManagerSpec {
 
     @Before
     fun setUp() {
-        doReturn(miniAppCustomPermission).whenever(miniApp).getCustomPermissions(miniAppId)
-        miniAppCustomPermissionManager = MiniAppCustomPermissionManager(miniApp)
+        doReturn(miniAppCustomPermission).whenever(miniAppCustomPermissionCache)
+            .readPermissions(miniAppId)
+        miniAppCustomPermissionManager = MiniAppCustomPermissionManager()
+    }
+
+    @Test
+    fun `preparePermissionsWithDescription should return a list of pair with name and description`() {
+        val description = "dummy description"
+        val customPermissionObj = CustomPermissionObj(
+            "rakuten.miniapp.user.USER_NAME",
+            description
+        )
+        val actual = miniAppCustomPermissionManager.preparePermissionsWithDescription(
+            arrayListOf(customPermissionObj)
+        )
+        val expected = listOf(Pair(MiniAppCustomPermissionType.USER_NAME, description))
+        assertEquals(expected, actual)
     }
 
     @Test
     fun `createJsonResponse should return json string based on supplied custom permissions`() {
         val miniAppId = "miniAppId"
         val suppliedPermissions = listOf(
-            Pair(MiniAppCustomPermissionType.USER_NAME, "description")
+            Pair(MiniAppCustomPermissionType.USER_NAME, "dummy description")
         )
-        val actual =
-            miniAppCustomPermissionManager.createJsonResponse(miniAppId, suppliedPermissions)
+        val actual = miniAppCustomPermissionManager.createJsonResponse(
+            miniAppCustomPermissionCache,
+            miniAppId,
+            suppliedPermissions
+        )
         val expected =
-            "{\"permissions\":[{\"name\":\"rakuten.miniapp.user.USER_NAME\",\"isGranted\":\"DENIED\"}]}"
+            "{\"permissions\":[{\"name\":\"rakuten.miniapp.user.USER_NAME\",\"status\":\"DENIED\"}]}"
 
         assertEquals(expected, actual)
     }
@@ -44,11 +62,14 @@ class MiniAppCustomPermissionManagerSpec {
     fun `filterPermissionsToSend should return correct values based on known custom permissions`() {
         val miniAppId = "miniAppId"
         val suppliedPermissions = listOf(
-            Pair(MiniAppCustomPermissionType.USER_NAME, "description")
+            Pair(MiniAppCustomPermissionType.USER_NAME, "dummy description")
         )
 
-        val actual =
-            miniAppCustomPermissionManager.filterPermissionsToSend(miniAppId, suppliedPermissions)
+        val actual = miniAppCustomPermissionManager.filterPermissionsToSend(
+            miniAppCustomPermissionCache,
+            miniAppId,
+            suppliedPermissions
+        )
         val expected = miniAppCustomPermission.pairValues
 
         assertEquals(expected, actual)
@@ -58,11 +79,14 @@ class MiniAppCustomPermissionManagerSpec {
     fun `filterPermissionsToSend should return correct values based on unknown custom permissions`() {
         val miniAppId = "miniAppId"
         val suppliedPermissions = listOf(
-            Pair(MiniAppCustomPermissionType.UNKNOWN, "description")
+            Pair(MiniAppCustomPermissionType.UNKNOWN, "dummy description")
         )
 
-        val actual =
-            miniAppCustomPermissionManager.filterPermissionsToSend(miniAppId, suppliedPermissions)
+        val actual = miniAppCustomPermissionManager.filterPermissionsToSend(
+            miniAppCustomPermissionCache,
+            miniAppId,
+            suppliedPermissions
+        )
         val expected = listOf(
             Pair(
                 MiniAppCustomPermissionType.UNKNOWN,